Revert "Updated code based on feedback"
This reverts commit f7b60ab883087a8453c93775d3da56b57c8bb41c.
