+++ prove/Develop03/Program.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 
 class Program
 {
+
     public static void Main()
     {
         Program program = new Program();
@@ -15,17 +15,8 @@ class Program
 
     public Program()
     {
-        // Initialize the scripture library
-        _scriptureLibrary = LoadScripturesFromFile("scriptures.txt");
-        if (_scriptureLibrary.Count > 0)
-        {
-            Random rand = new Random();
-            _scripture = _scriptureLibrary[rand.Next(_scriptureLibrary.Count)];
-        }
-        else
-        {
-            Console.WriteLine("No scriptures loaded.");
-        }
+        // Initialize the scripture
+        InitializeScriptures();
     }
 
     public void Run()
@@ -46,6 +37,7 @@ class Program
 
             _scripture.HideWords();
 
+            // Check again after hiding words to ensure the final state is displayed
             if (_scripture.IsCompletelyHidden())
             {
                 Console.Clear();
@@ -57,18 +49,51 @@ class Program
         Console.WriteLine("\nMemorization complete! Well done!");
     }
 
-    private List<Scripture> LoadScripturesFromFile(string filePath)
+
+    private void InitializeScriptures()
     {
-        var scriptures = new List<Scripture>();
+        _scriptureLibrary = new List<Scripture>();
+
+        // Replace with the path to your text file
+        string filePath = "scriptures.txt";
 
         if (File.Exists(filePath))
         {
             foreach (var line in File.ReadLines(filePath))
             {
-                var scripture = Scripture.ParseFromLine(line);
-                if (scripture != null)
+                // Split the line into reference and text based on " - "
+                var parts = line.Split(" - ");
+                if (parts.Length == 2)
                 {
-                    scriptures.Add(scripture);
+                    var referenceText = parts[0]; // e.g., "Hebrews 11:1"
+                    var scriptureText = parts[1].Trim('"'); // Remove the quotes around the text
+
+                    // Parse the reference
+                    var referenceParts = referenceText.Split(':');
+                    if (referenceParts.Length == 2)
+                    {
+                        var bookAndChapter = referenceParts[0].Trim();
+                        var verseRange = referenceParts[1].Trim();
+
+                        var bookAndChapterParts = bookAndChapter.Split(' ');
+                        if (bookAndChapterParts.Length == 2)
+                        {
+                            string book = bookAndChapterParts[0];
+                            int chapter = int.Parse(bookAndChapterParts[1]);
+   
+                            // Handle possible range in verse (e.g., "1-4" or "1")
+                            var verses = verseRange.Split('-');
+                            int startVerse = int.Parse(verses[0]);
+                            int endVerse = verses.Length == 2 ? int.Parse(verses[1]) : startVerse;
+
+                            // Create the Reference object
+                            var reference = new Reference(book, chapter, startVerse, endVerse);
+
+                            // Create the Scripture object and add it to the list
+                            var scripture = new Scripture(reference, scriptureText);
+                            _scriptureLibrary.Add(scripture);
+                        }
+                    }
                 }
             }
         }
@@ -77,7 +102,12 @@ class Program
             Console.WriteLine("Scripture file not found.");
         }
 
-        return scriptures;
+        // Select a random scripture from the list
+        if (_scriptureLibrary.Count > 0)
+        {
+            Random rand = new Random();
+            _scripture = _scriptureLibrary[rand.Next(_scriptureLibrary.Count)];
+        }
     }
 
     private void DisplayOptions()
